requirements.txt
Django==1.10.5
django-ckeditor==5.2.1
django-grappelli==2.9.1
Pillow==4.0.0
pytz==2016.10
